// src/app/routes/recepcion/inventario/interfaces/response-get-caja-interface.ts
// Generated by https://quicktype.io

export interface ResponseGetCaja {
  id: number;
  numEmpaque: string;
  numEnvio: string;
  estado: Estado;
  listDocumento: ListDocumento[];
  coAnaquel: string;
  coLado: string;
  coFila: string;
  coColumna: string;
}

export interface Estado {
  codigo: string;
  descripcion: string;
}

export interface ListDocumento {
  id: number;
  tipoDocu: AreaOrigen;
  areaOrigen: AreaOrigen;
  numSerieDocu: string;
  numPaquete: string;
  numFolios: number;
  estado: Estado;
}

export interface AreaOrigen {
  codigo: string;
}